+++ src/glicko.rs
@@ -21,9 +21,9 @@ impl Rating {
         self.rd * self.rd
     }
 
+    /// This assumes 30 2 month periods must pass before one's rating
+    /// deviation is the same as a new player and that a typical RD is 50.
     pub fn update_rd(&mut self) {
-        // This assumes 30 2 month periods must pass before one's rating
-        // deviation is the same as a new player and that a typical RD is 50.
         let c = 63.2;
 
         let rd_new = f64::sqrt(self.rd_sq() + (c * c));