+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Run update_rd every two months.
 - You're allowed to resign.
 - Displays when the game is over.
 - The option to leave, accept, or decline a game.